DTO, Entity 추가

+++ src/main/java/com/ohdaesan/shallwepets/chattingRoom/repository/ChattingRoomRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChattingRoomRepository extends JpaRepository<ChattingRoom, Integer> {
+public interface ChattingRoomRepository extends JpaRepository<ChattingRoom, Long> {
 }
