+++ src/main/java/com/ohdaesan/shallwepets/images/repository/ImagesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImagesRepository extends JpaRepository<Images, Integer> {
+public interface ImagesRepository extends JpaRepository<Images, Long> {
 }
